+++ examples/advanced/spline_draw.py
@@ -49,7 +49,7 @@ outfl = 'spline.txt'
 cpoints = []
 points, spline= None, None
 
-pic = Picture("https://embryology.med.unsw.edu.au/embryology/images/4/40/Mouse-_embryo_E11.5.jpg",
+pic = Picture(dataurl+"images/Mouse-_embryo_E11.5.jpg",
               channels=[0,1,2]) # keep rgb but drop alpha channel
 
 t = """Click to add a point