+++ examples/advanced/warp5.py
@@ -1,143 +1,148 @@
-# Morph one shape into another interactively (can work in 3d too!)
-#
-from vedo import Plotter, Axes, dataurl, load, printc, merge
-from vedo.shapes import Text2D, Points, Lines, Arrows2D, Grid
+"""
+Takes 2 shapes, source and target, and morphs source on target
+this is obtained by fitting 18 parameters of a non linear,
+quadratic, transformation defined in transform()
+The fitting minimizes the distance to the target surface
+using algorithms available in the scipy.optimize package.
+"""
+from vedo import *
+import scipy.optimize as opt
+
+print(__doc__)
+
+settings.useDepthPeeling = True
+
+plt = Plotter(shape=[1, 3], interactive=0, axes=1)
+
 
 class Morpher:
+    def __init__(self):
+        self.source = None
+        self.target = None
+        self.bound = 0.1
+        self.method = "SLSQP"  # 'SLSQP', 'L-BFGS-B', 'TNC' ...
+        self.tolerance = 0.0001
+        self.subsample = 200  # pick only subsample pts
+        self.allowScaling = False
+        self.params = []
 
-    def __init__(self, mesh1, mesh2, n):  ############################### init
+        self.msource = None
+        self.s_size = ([0, 0, 0], 1)  # ave position and ave size
+        self.fitResult = None
+        self.chi2 = 1.0e10
 
-        self.n = n  # desired nr. of intermediate shapes
-        self.mode = '2d'
-        self.mesh1 = mesh1
-        self.mesh2 = mesh2
-        self.merged_meshes = merge(mesh1, mesh2)
-        self.mesh1.lw(4).c('grey2').pickable(False)
-        self.mesh2.lw(4).c('grey1').pickable(False)
+    # -------------------------------------------------------- fit function
+    def transform(self, p):
+        a1, a2, a3, a4, a5, a6, b1, b2, b3, b4, b5, b6, c1, c2, c3, c4, c5, c6, s = self.params
+        pos, sz = self.s_size[0], self.s_size[1]
+        x, y, z = (p - pos) / sz * s  # bring to origin, norm and scale
+        xx, yy, zz, xy, yz, xz = x * x, y * y, z * z, x * y, y * z, x * z
+        xp = x + 2 * a1 * xy + a4 * xx + 2 * a2 * yz + a5 * yy + 2 * a3 * xz + a6 * zz
+        yp = +2 * b1 * xy + b4 * xx + y + 2 * b2 * yz + b5 * yy + 2 * b3 * xz + b6 * zz
+        zp = +2 * c1 * xy + c4 * xx + 2 * c2 * yz + c5 * yy + z + 2 * c3 * xz + c6 * zz
+        p2 = vector(xp, yp, zp)
+        p2 = (p2 * sz) + pos  # take back to original size and position
+        return p2
 
-        self.arrow_starts = []
-        self.arrow_stops  = []
-        self.dottedln = None
-        self.toggle = False
-        self.instructions = ("Click to add arrows interactively on the left panel\n"
-                             "right-click to remove last arrow. Then press:\n"
-                             "- m to morph the plane\n"
-                             "- c to clear\n"
-                             "- g to generate interpolation")
-        self.msg1 = Text2D(self.instructions, pos='top-left', font="VictorMono", bg='g2', alpha=0.6)
-        self.msg2 = Text2D('[output will show here]', pos='top-left', font="VictorMono")
+    def _func(self, pars):
+        self.params = pars
 
-        sz = self.merged_meshes.diagonalSize()
-        self.plane1 = Grid(sx=sz, sy=sz, resx=50, resy=50).pos(self.merged_meshes.centerOfMass())
-        self.plane1.wireframe(False).alpha(1).lineWidth(0.1).c('white').lc('grey5')
-        self.plane2 = self.plane1.clone().pickable(False)
+        #calculate chi2
+        d2sum, n = 0.0, self.source.N()
+        srcpts = self.source.points()
+        rng = range(0, n, int(n / self.subsample))
+        for i in rng:
+            p1 = srcpts[i]
+            p2 = self.transform(p1)
+            tp = self.target.closestPoint(p2)
+            d2sum += mag2(p2 - tp)
+        d2sum /= len(rng)
 
-        self.plotter = Plotter(N=2, bg='light blue', size=(2000,1000))
-        self.plotter.addCallback('left click', self.onleftclick)
-        self.plotter.addCallback('right click', self.onrightclick)
-        self.plotter.addCallback('key press', self.onkeypress)
+        if d2sum < self.chi2:
+            if d2sum < self.chi2 * 0.99:
+                print("Emin ->", d2sum)
+            self.chi2 = d2sum
+        return d2sum
 
-    def start(self):  ################################################ show stuff
-        paxes = Axes(self.plane1, xyGrid=0, textScale=0.6)
-        self.plotter.show(self.plane1, paxes, self.msg1, self.mesh1, self.mesh2, at=0)
-        self.plotter.show(self.plane2, self.msg2, at=1, mode='image')
-        if len(self.arrow_starts):
-            self.draw(True)
-            self.draw(False)
-            self.msg1.text(self.instructions)
-        self.plotter.show(interactive=True, zoom=1.3).close()
+    # ------------------------------------------------------- Fit
+    def morph(self):
+        def avesize(pts):  # helper fnc
+            s, amean = 0, vector(0, 0, 0)
+            for p in pts:
+                amean = amean + p
+            amean /= len(pts)
+            for p in pts:
+                s += mag(p - amean)
+            return amean, s / len(pts)
 
-    def draw(self, toggle=None):  #################################### update scene
-        if toggle is None:
-            toggle = self.toggle
-        if toggle:
-            self.msg1.text("Choose start point or press:\nm to morph the shapes\ng to interpolate")
-            self.plotter.remove("displacementArrows", at=0)
-            if not len(self.arrow_starts): return
-            arrows = Arrows2D(self.arrow_starts, self.arrow_stops, c='red4')
-            arrows.name = "displacementArrows"
-            self.plotter.add(arrows, at=0)
+        print("\n..minimizing with " + self.method)
+        self.msource = self.source.clone()
+
+        self.s_size = avesize(self.source.points())
+        bnds = [(-self.bound, self.bound)] * 18
+        x0 = [0.0] * 18  # initial guess
+        x0 += [1.0]  # the optional scale
+        if self.allowScaling:
+            bnds += [(1.0 - self.bound, 1.0 + self.bound)]
         else:
-            self.msg1.text("Click to choose an end point")
-            self.plotter.remove("displacementPoints", at=0)
-            points = Points(self.arrow_starts, r=15, c='green3', alpha=0.5)
-            points.name = "displacementPoints"
-            self.plotter.add(points, at=0)
+            bnds += [(1.0, 1.0)]  # fix scale to 1
+        res = opt.minimize(self._func, x0,
+                           bounds=bnds, method=self.method, tol=self.tolerance)
+        # recalc for all pts:
+        self.subsample = self.source.N()
+        self._func(res["x"])
+        print("\nFinal fit score", res["fun"])
+        self.fitResult = res
 
-    def onleftclick(self, evt):  ############################################ add points
-        msh = evt.actor
-        if not msh or msh.name!="Grid": return
-        pt = self.merged_meshes.closestPoint(evt.picked3d) # get the closest pt on the line
-        self.arrow_stops.append(pt) if self.toggle else self.arrow_starts.append(pt)
-        self.draw()
-        self.toggle = not self.toggle
+    # ------------------------------------------------------- Visualization
+    def draw_shapes(self):
 
-    def onrightclick(self, evt):  ######################################## remove points
-        if not self.arrow_starts: return
-        self.arrow_starts.pop()
-        if not self.toggle:
-            self.arrow_stops.pop()
-        self.plotter.clear(at=0).addRendererFrame()
-        self.plotter.add([self.plane1, self.msg1, self.mesh1, self.mesh2], at=0, render=False)
-        self.draw(False)
-        self.draw(True)
+        pos, sz = self.s_size[0], self.s_size[1]
 
-    def onkeypress(self, evt):  ###################################### MORPH & GENERATE
-        if evt.keyPressed == 'm': ##--------- morph mesh1 based on the existing arrows
-            if len(self.arrow_starts) != len(self.arrow_stops):
-                printc("You must select your end point first!", c='y')
-                return
+        sphere0 = Sphere(pos, c="gray", r=sz, alpha=0.8, res=16).wireframe()
+        sphere1 = sphere0.clone().alpha(0.2).wireframe(False)
 
-            output = [self.mesh1.clone().c('grey4'), self.mesh2, self.msg2]
-            warped_plane = self.plane1.clone().pickable(False)
-            warped_plane.warp(self.arrow_starts, self.arrow_stops, mode=self.mode)
-            output.append(warped_plane + Axes(warped_plane, xyGrid=0, textScale=0.6))
+        newpts = []
+        for p in self.msource.points():
+            newp = self.transform(p)
+            newpts.append(newp)
+        self.msource.points(newpts)
 
-            mw = self.mesh1.clone().applyTransform(warped_plane.transform).c('red4')
-            output.append(mw)
+        arrs = []
+        newpts = []
+        for p in sphere0.points():
+            newp = self.transform(p)
+            newpts.append(newp)
+            arrs.append([p, newp])
+        sphere1.points(newpts)
+        hair = Arrows(arrs, s=0.3, alpha=0.5, c='jet')
 
-            T_inv = warped_plane.transform.GetInverse()
-            a = Points(self.arrow_starts, r=10).applyTransform(warped_plane.transform)
-            b = Points(self.arrow_stops,  r=10).applyTransform(warped_plane.transform)
-            self.dottedln = Lines(a,b, res=self.n).applyTransform(T_inv).pointSize(5)
-            output.append(self.dottedln)
+        zero = Point(pos, c="black")
+        x1, x2, y1, y2, z1, z2 = self.target.polydata().GetBounds()
+        tpos = [x1, y2, z1]
+        text1 = Text3D("source vs target", tpos, s=sz / 10, c="dg")
+        text2 = Text3D("morphed vs target", tpos, s=sz / 10, c="dg")
+        text3 = Text3D("deformation", tpos, s=sz / 10, c="dr")
 
-            self.msg1.text(self.instructions)
-            self.msg2.text("Morphed output:")
-            self.plotter.clear(at=1).addRendererFrame().add(output, at=1)
-
-        elif evt.keyPressed == 'g':  ##------- generate intermediate shapes
-            if not self.dottedln:
-                return
-            intermediates = []
-            allpts = self.dottedln.points()
-            allpts = allpts.reshape(len(self.arrow_starts), self.n+1, 3)
-            for i in range(self.n + 1):
-                pi = allpts[:,i,:]
-                m_nterp = self.mesh1.clone().warp(self.arrow_starts, pi, mode=self.mode).c('b3').lw(1)
-                intermediates.append(m_nterp)
-            self.msg2.text("Morphed output + Interpolation:")
-            self.plotter.add(intermediates, at=1)
-            self.dottedln = None
-
-        elif evt.keyPressed == 'c':  ##------- clear all
-            self.arrow_starts = []
-            self.arrow_stops  = []
-            self.toggle = False
-            self.dottedln = None
-            self.msg1.text(self.instructions)
-            self.msg2.text("[output will show here]")
-            self.plotter.clear(at=0).clear(at=1).addRendererFrame()
-            self.plotter.add([self.plane1, self.msg1, self.mesh1, self.mesh2], at=0)
-            self.plotter.add([self.plane2, self.msg2], at=1)
+        plt.show(sphere0, sphere1, zero, text3, hair, at=2)
+        plt.show(self.msource, self.target, text2, at=1)
+        plt.show(self.source, self.target, text1, at=0, zoom=1.2, interactive=1)
+        plt.close()
 
 
-######################################################################################## MAIN
+#################################
 if __name__ == "__main__":
-    outlines = load(dataurl+"timecourse1d.npy") # load a set of 2d shapes
-    mesh1 = outlines[25]
-    mesh2 = outlines[35].scale(1.3).shift(-2,0,0)
-    morpher = Morpher(mesh1, mesh2, 10)  # generate 10 intermediate outlines
-    morpher.start()
 
+    mr = Morpher()
+    mr.source = plt.load(dataurl+"270.vtk").color("g").alpha(0.4)
+    mr.target = plt.load(dataurl+"290.vtk").color("b").alpha(0.3)
+    mr.target.wireframe()
+    mr.allowScaling = True
+    mr.bound = 0.4  # limits the parameter value
 
+    mr.morph()
+
+    print("Result of parameter fit:\n", mr.params)
+
+    # now mr.msource contains the modified/morphed source.
+    mr.draw_shapes()